refactor: store syncing mode in its own file
