+++ packages/loot-core/src/server/sync/index.js
@@ -1,6 +1,4 @@
 export { 
-  checkSyncingMode,
-  setSyncingMode,
   serializeValue,
   deserializeValue,
   addSyncListener,
@@ -15,6 +13,10 @@ export {
   initialFullSync,
   fullSync
 } from "./sync"
+export { 
+  checkSyncingMode,
+  setSyncingMode,
+} from "./syncing-mode"
 export { default as makeTestMessage } from './make-test-message';
 export { default as resetSync } from './reset';
 export { default as repairSync } from './repair';
